Rename "posts" package to "content".
This is in preparation to write the "About" page. For
simplicity's sake, the "About" page will have the same
structure as a blog post - both as the frontend will see it
and as the backend will use it. For this reason, we should
reuse the code already written for the "posts" pkg, but the name
"posts" is no longer accurate. Rather the name "content"
is broader, and includes the idea of both an "about" page
and the "posts" pages.